Prevent the ship from flying off the screen
I tried to do this a few ways, and this is the one that worked best. The problem
with registering a simple `Outside Room` event like one would expect is that (I
think) the `<Left>` event keeps firing, so we toggle the ship's speed between 0
and 10, and the result is it keeps moving. We could avoid continuing to send
`<Left>`s by checking a `Key Press` instead of a `Keyboard`, but that leads to
funky controls that don't behave the way you'd expect as you fly around.

So though it seems a little clumsy, this approach with the `if`/`else` seems to
work best. Whenever `<Left>` fires, we check if the ship is capable of moving
left (`x >= 10`) or right (`x <= (640 - 32)`) and set the speed to either 10 or
0.

`galaga-final.gmk` contains my progress through the lesson, while `galaga.gmk`
will remain the initial state, which is bundled up in the `zip`.