Add Win and Lose states
When your ship explodes, "Game Over" gets printed. I tried using "Draw Text",
but it was finicky and only drew the plainest text ever.

When you shoot the last baddie, we notice via a "Test Instance Count" and print
"You Win!"

I refactored the pieces responsible for various destructive tendencies. Now when
a baddie goes up in flames, its own `Destroy` event is responsible for creating
the explosion and updating the score. Both missiles is now only responsible for
destroying themself and the thing they hit. That way, we can also have the bad
missile create the Game Over state when it hits the ship, and vise versa for the
good missile and baddies.